+++ src/Locator/AbstractResultSet.php
@@ -11,4 +11,13 @@ use Dhii\SimpleTest\Test;
  */
 abstract class AbstractResultSet extends Test\AbstractSource implements ResultSetInterface
 {
+    /**
+     * Makes the set unique by identifying each item by a key derived from the item.
+     *
+     * @since 0.1.1
+     */
+    protected function _getItemUniqueKey($item)
+    {
+        return $this->_getItemKey($item);
+    }
 }