+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change log
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/) 
+The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.1.0 - 2016-11-10
 The initial release
+
+## 0.1.1 - 2016-11-15
+Quick release to fix some bugs.
+
+### Changed
+- `Locator\AbstractResultSet` items will now be unique, and their keys preserved and exposed in iteration.
+- Tests slightly improved.